Modif + Supp Profil OK

+++ frontend/src/services/user.js
@@ -18,7 +18,7 @@ class UserService {
         return api.put( '/user/' + id , data, { headers: auth() })
     }
     deleteUser(id) {
-        return api.delete('/user/' + id, { headers: auth() });
+        return api.delete('/user/' + id, { headers: auth(), data: { userId: id }});
     }
 }
 export default new UserService();
